+++ src/UseContext.js
@@ -3,6 +3,7 @@ import Booklist from "./components/useContextComponents/Booklist";
 import Navbar from "./components/useContextComponents/Navbar";
 import ThemeToggle from "./components/useContextComponents/ThemeToggle";
 import AuthContextProvider from "./contexts/AuthContext";
+import BookContextProvider from "./contexts/BookContext";
 import ThemeContextProvider from "./contexts/ThemeContext";
 
 class Context extends Component {
@@ -12,7 +13,9 @@ class Context extends Component {
 				<ThemeContextProvider>
 					<AuthContextProvider>
 						<Navbar />
-						<Booklist />
+                        <BookContextProvider>
+						    <Booklist />
+                        </BookContextProvider>
 						<ThemeToggle />
 					</AuthContextProvider>
         </ThemeContextProvider>
